Simplify PMTiles platform source sets
