--- pmtiles/src/jsMain/kotlin/org/maplibre/spatialk/pmtiles/internal/GzipJs.kt
@@ -1,4 +1,0 @@
-package org.maplibre.spatialk.pmtiles.internal
-
-internal actual fun decodeGzip(bytes: ByteArray, limits: DecodeLimits): ByteArray =
-    throw NotImplementedError("gzip decoding is not implemented for Kotlin/JS.")
