--- pmtiles/src/jsTest/kotlin/org/maplibre/spatialk/pmtiles/internal/GzipJsTest.kt
@@ -1,14 +1,0 @@
-package org.maplibre.spatialk.pmtiles.internal
-
-import kotlin.test.Test
-import kotlin.test.assertFailsWith
-import org.maplibre.spatialk.pmtiles.Compression
-
-class GzipJsTest {
-    @Test
-    fun gzipPlaceholderThrowsWhenInvoked() {
-        assertFailsWith<NotImplementedError> {
-            decodeCompression(Compression.Gzip, helloGzipBytes, testDecodeLimits())
-        }
-    }
-}
