+++ pmtiles/src/notImplementedMain/kotlin/org/maplibre/spatialk/pmtiles/internal/GzipNotImplemented.kt
@@ -1,4 +1,4 @@
 package org.maplibre.spatialk.pmtiles.internal
 
 internal actual fun decodeGzip(bytes: ByteArray, limits: DecodeLimits): ByteArray =
-    throw NotImplementedError("gzip decoding is not implemented for Kotlin/Wasm JS.")
+    throw NotImplementedError("gzip decoding is not implemented for this Kotlin target.")
