--- pmtiles/src/wasmWasiMain/kotlin/org/maplibre/spatialk/pmtiles/internal/GzipWasmWasi.kt
@@ -1,4 +1,0 @@
-package org.maplibre.spatialk.pmtiles.internal
-
-internal actual fun decodeGzip(bytes: ByteArray, limits: DecodeLimits): ByteArray =
-    throw NotImplementedError("gzip decoding is not implemented for Kotlin/Wasm WASI.")
